fixed log in and sign up working great now and getting username in profile
