+++ config/passport.js
@@ -6,21 +6,19 @@ var db = require("../models");
 
 // Telling passport we want to use a Local Strategy. In other words, we want login with a username/email and password
 passport.use(new LocalStrategy(
-    // Our user will sign in using an email, rather than a "username"
-    {
-        usernameField: "email"
-    },
-    function (email, password, done) {
+    function (username, password, done) {
         // When a user tries to sign in this code runs
+        console.log("----------about to check database for username-----------")
         db.User.findOne({
             where: {
-                email: email
+                username: username
             }
         }).then(function (dbUser) {
-            // If there's no user with the given email
+            console.log("----------- checked database -----------")
+            // If there's no user with the given username
             if (!dbUser) {
                 return done(null, false, {
-                    message: "Incorrect email."
+                    message: "Incorrect username."
                 });
             }
             // If there is a user with the given email, but the password the user gives us is incorrect
