+++ routes/api-routes.js
@@ -11,17 +11,25 @@ var passport = require("../config/passport");
 module.exports = function (app) {
   // code from 14-homework for validating the hashed password
   app.post("/api/login", passport.authenticate("local"), function (req, res) {
-    res.json(req.User);
+    res.json(req.user);
   });
-
+  // new code from the docs for passport
+  // app.post('/login',
+  //   passport.authenticate('local', {
+  //     successRedirect: '/profile',
+  //     failureRedirect: '/',
+  //     failureFlash: true
+  //   })
+  // );
   // Route for signing up a user. The user's password is automatically hashed and stored securely thanks to
   // how we configured our Sequelize User Model. If the user is created successfully, proceed to log the user in,
   // otherwise send back an error
   app.post("/api/signup", function (req, res) {
     console.log(req.body);
+    console.log("-----------about to send to DB---------------")
     db.User.create({
       name: req.body.name,
-      userName: req.body.userName,
+      username: req.body.username,
       password: req.body.password
     })
       .then(function () {
@@ -41,17 +49,18 @@ module.exports = function (app) {
   });
 
   // Route for getting some data about our user to be used client side
-  app.get("/api/userprofile", function (req, res) {
-    if (!req.User) {
+  app.get("/api/user_data", function (req, res) {
+    if (!req.user) {
+      console.log("usernot logged in so no data passed")
       // The user is not logged in, send back an empty object
       res.json({});
     } else {
-
+      console.log("user logged in pass the data")
       // Otherwise send back the user's email and id
       // Sending back a password, even a hashed password, isn't a good idea
       res.json({
-        userName: req.User.userName,
-        id: req.User.id
+        username: req.user.username,
+        id: req.user.id
       });
     }
   });
