+++ views/index.handlebars
@@ -31,7 +31,7 @@
                     <img class="login-logo" src="../images/icons/blackIcons/logoBK.png" alt="Avatar" class="avatar">
                 </div>
 
-                <form id="sign-in-form" class="form">
+                <form id="sign-in-form" class="form" action="/login" method="post">
                     <label class="lbl" for='user-name'><span class="required">* </span>Username</label>
                     <input class="lbl" id='login-username' type='text' name='username'>
                     <label class="lbl" for='password'><span class="required">* </span>Password</label>
