+++ views/layouts/main.handlebars
@@ -18,6 +18,8 @@
     {{{ body }}}
     {{!-- custom scripts --}}
     <script src="assets/js/index.js"></script>
+    <script src="assets/js/login.js"></script>
+    <script src="assets/js/signUp.js"></script>
 </body>
 
 </html>
