+++ views/profile.handlebars
@@ -10,7 +10,7 @@
 {{!-- main profile container --}}
 <div class="container center" style="padding: 100px 15px;">
     {{!-- putting placeholder text for the Username --}}
-    <h3 id="dad-username" class="center" style="padding-bottom: 15px;">Dadbod21</h3>
+    <h3 id="dad-username" class="center" style="padding-bottom: 15px;"></h3>
 
     <div class="dadPoints">
         {{!-- this is just placeholder till the DB is working --}}
@@ -25,7 +25,6 @@
 
     <!--dad points with span element on unique number-->
     <h4 class="post-a-moment">Post a joke, Share a dad moment</h4>
-    <p style="display: none;" class="error-message">Please make sure you have up to 100 charactrs</p>
     <form class="moment-form">
         <textarea type='text' id='dadmmnt' name='moments' minlength="25"></textarea><br><br>
         <input type="submit" class='btn btn-primary profile-submit'></input><br><br>
@@ -81,6 +80,14 @@
             })
 
         });
+        // adding the dad username
+
+        // This file just does a GET request to figure out which user is logged in
+        // and updates the HTML on the page
+        $.get("/api/user_data").then(function (data) {
+            dadUsername.text(data.username);
+        });
+
     });
